move customer handler and its tests
